//some changes added to saga and outbox model infrastructure

+++ commonService/src/main/kotlin/com/micro/commonservice/Order.kt
@@ -1,10 +1,12 @@
 package com.micro.commonservice
 
+import java.math.BigDecimal
+
 data class Order(
     val id: Long,
     val customerId: Long,
     val productId: Long,
-    val price: Int,
+    val price: BigDecimal,
     var status: String,
     val productCount: Int
     // other properties ....
